Skip comm_world list: Virtualize all access to the VCR list
    Also, use "pami_task_t" in the rank list instead of "int".

+++ mpich2/src/mpid/pamid/src/mpid_vc.c
@@ -29,7 +29,7 @@ int MPID_VCR_Dup(MPID_VCR orig_vcr, MPID_VCR * new_vcr)
 
 int MPID_VCR_Get_lpid(MPID_VCR vcr, int * lpid_ptr)
 {
-    *lpid_ptr = vcr;
+    *lpid_ptr = (int)vcr;
     return MPI_SUCCESS;
 }
